requirements.txt
# Обязательные зависимости
fpdf2>=2.7.0